requirements.txt
boto3
fastapi
uvicorn
pydantic
python-jose
python-multipart
mangum
python-dotenv